chore(automations): move bespoke SerializedToJson field type to automations
