+++ wandb/_pydantic/__init__.py
@@ -1,7 +1,7 @@
 """Internal utilities for working with pydantic."""
 
 from .base import CompatBaseModel, GQLBase
-from .field_types import GQLId, SerializedToJson, Typename, ensure_json
+from .field_types import GQLId, Typename
 from .utils import IS_PYDANTIC_V2, from_json, gql_typename, pydantic_isinstance, to_json
 from .v1_compat import (
     AliasChoices,
@@ -17,7 +17,6 @@ __all__ = [
     "GQLBase",
     "Typename",
     "GQLId",
-    "SerializedToJson",
     "AliasChoices",
     "computed_field",
     "field_validator",
@@ -26,6 +25,5 @@ __all__ = [
     "to_camel",
     "to_json",
     "from_json",
-    "ensure_json",
     "gql_typename",
 ]
